move blog posts to index page
