+++ pages/index.js
@@ -1,20 +1,9 @@
 import Head from 'next/head'
 import Link from 'next/link'
 import Layout, { siteTitle } from '../components/layout'
-import Resume from './resume'
 import utilStyles from '../styles/utils.module.css'
-import { getSortedPostData } from '../lib/posts'
 
-export async function getStaticProps() {
-  const allPostsData = getSortedPostData()
-  return {
-    props: {
-      allPostsData
-    }
-  }
-}
-
-export default function Home({ allPostsData }) {
+export default function Home() {
   return (
     <Layout home>
       <Head>
@@ -28,22 +17,6 @@ export default function Home({ allPostsData }) {
         </p>
       </section>
       <Link href="/resume"><a>View Resume</a></Link>
-      <section className={`${utilStyles.headingMd} ${utilStyles.padding1px}`}>
-        <h2 className={utilStyles.headingLg}>Blog</h2>
-        <ul className={utilStyles.list}>
-          {allPostsData.map(({ id, date, title }) => (
-            <li className={utilStyles.listItem} key={id}>
-              <Link href={`/posts/${id}`}>
-                <a>{title}</a>
-              </Link>
-              <br />
-              <small className={utilStyles.lightText}>
-                <Date dateString={date} />
-              </small>
-            </li>
-          ))}
-        </ul>
-      </section>
     </Layout>
   )
 }
