+++ pages/posts/index.js
@@ -1,0 +1,38 @@
+import Layout from '../../components/layout';
+import utilStyles from '../../styles/utils.module.css';
+import { getSortedPostData } from '../../lib/posts';
+import Link from 'next/link';
+
+
+export async function getStaticProps() {
+  const allPostsData = getSortedPostData()
+  return {
+    props: {
+      allPostsData
+    }
+  }
+}
+
+export default function PostsLanding({ allPostsData }) {
+  return (
+    <Layout>
+      <h1>Here's all the posts</h1>
+      <section className={`${utilStyles.headingMd} ${utilStyles.padding1px}`}>
+        <h2 className={utilStyles.headingLg}>Blog</h2>
+        <ul className={utilStyles.list}>
+          {allPostsData.map(({ id, date, title }) => (
+            <li className={utilStyles.listItem} key={id}>
+              <Link href={`/posts/${id}`}>
+                <a>{title}</a>
+              </Link>
+              <br />
+              <small className={utilStyles.lightText}>
+                <Date dateString={date} />
+              </small>
+            </li>
+          ))}
+        </ul>
+      </section>
+    </Layout>
+  )
+}
